requirements.txt
fastapi>=0.109.0
uvicorn>=0.23.2
motor>=3.1.2
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.1
langchain-chroma>=0.0.1
openai>=1.0.0
chromadb>=0.4.6
sentence-transformers>=2.2.2
google-cloud-speech>=2.19.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.0.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
email-validator>=2.0.0
google-auth>=2.22.0
google-generativeai>=0.3.1
pinecone-client>=6.0.0